Removing items from inventory now works
You can now drag items out of the inventory and it will set the item
prefab active again and change its position to the mouse position.
Also there was a bug when the inventory was closed, a bunch of bullets
would be fired, so I added a timescale == 1 condition when clicking in
the Player script.